Learner/requirements.txt
scikit-learn
kafka-python
numpy
pandas
pytest